Ecommerce consumer behaviour dashboard/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
duckdb==1.1.3
pandas==2.3.2
numpy==2.1.1
scikit-learn==1.5.2
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
pytest==8.4.1
httpx==0.28.1
slowapi==0.1.9
python-jose[cryptography]==3.3.0
